Longest Substring Without Repeating Characters
Double pointers method. The front pointer runs from the left to the right. Use a hash map to record existing characters. If a character is found by the front pointer, do the following two things: 1. update the back pointer by the original index adding 1; 2, update the recorded index by the newly found index; 3 update the maximal length. This method is a kind of dynamic programming: if the newly found character is same as one in the previous string, all characters before the previous character should be discarded.